requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.27.2=pypi_0
astunparse=1.6.3=pypi_0
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.7.4=pypi_0
charset-normalizer=3.3.2=pypi_0
filelock=3.15.4=pypi_0
fsspec=2024.6.1=pypi_0
huggingface-hub=0.24.5=pypi_0
idna=3.7=pypi_0
importlib-metadata=8.2.0=pypi_0
jdiffusion=1.0=dev_0
jittor=1.3.9.5=dev_0
jtorch=0.1.8=dev_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.6=pypi_0
markupsafe=2.1.5=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.26.4=pypi_0
openssl=3.0.14=h5eee18b_0
packaging=24.1=pypi_0
peft=0.10.0=pypi_0
pillow=10.4.0=pypi_0
pip=24.0=py39h06a4308_0
protobuf=5.27.3=pypi_0
psutil=6.0.0=pypi_0
python=3.9.19=h955ad1f_1
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
regex=2024.7.24=pypi_0
requests=2.32.3=pypi_0
safetensors=0.4.4=pypi_0
setuptools=72.1.0=py39h06a4308_0
six=1.16.0=pypi_0
sqlite=3.45.3=h5eee18b_0
tensorboard=1.10.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.13.3=pypi_0
torch=2.0.0=pypi_0
tqdm=4.66.5=pypi_0
transformers=4.26.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py39h06a4308_0
xz=5.4.6=h5eee18b_1
zipp=3.20.0=pypi_0
zlib=1.2.13=h5eee18b_1
